메인.js.파일에 있는 함수들 다른 파일로 빼서 import시켯슴다. 메인화면 페이드인아웃 구현햇슴다.
