+++ src/lib/util.js
@@ -1,13 +1,12 @@
 const _ = {
-    ajaxFunc : function ajax(url, exeFunc) {
+    ajaxFunc : function ajax(url, exeFunc, selector) {
     const oReq = new XMLHttpRequest(); 
         oReq.addEventListener("load", (e) => {
             const data = JSON.parse(oReq.responseText);
-            exeFunc();
+            exeFunc(data, selector);
         }); 
         oReq.open("GET", url); 
         oReq.send();
     }
 }
-
-
+export default _;
