+++ src/main.html
@@ -8,6 +8,9 @@
     <body>
         <div class="wrap">
             <header class="main_header">
+                <div class="main_header_fade_in"></div>
+                <div class="main_header_fade_out"></div>
+                <div class="main_header_dimmed"></div>
                 <nav class="head_menu">
                     <div class="logo_wrap">
                         <div class="fling_logo"></div>
@@ -35,7 +38,7 @@
             
 
                 <div class="search_desc">
-                    '해먹남녀'에서 원하는 레시피링크를 복사, 붙여넣기 해주세요.
+                    '해먹남녀'에서 원하는 레시피 링크를 복사, 붙여넣기 해주세요.
                 </div>
                 <div class="recipe_site_link">
                     <a target="_blank" href="http://haemukja.com/">
@@ -224,6 +227,6 @@
             </div>
         </div>
         <script src="https://cdnjs.cloudflare.com/ajax/libs/handlebars.js/4.0.10/handlebars.js"></script>
-        <script src="./main.js"></script>
+        <script src="../dist/bundle.js"></script>
     </body>
 </html>
